final-pjt-back/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==4.2.13
django-allauth==0.63.0
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.1
idna==3.7
importlib_metadata==7.1.0
Markdown==3.6
oauthlib==3.2.2
pillow==10.3.0
pycparser==2.22
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
zipp==3.18.2
